+++ src/components/VoteExplainerCard/VoteExplainerCard.tsx
